sircleORA

Uses enricher to run ORA on each of the clusters as defined by the regulatory labels using prior knowledge such as gene-sets

Usage

sircleORA(
  InputData,
  geneID,
  regLabels = "RegulatoryLabels",
  PriorKnowledge,
  PKName = "",
  emptyRegLabel = "",
  RemoveBackgroundGenes = "TRUE",
  Plot_FileType = "pdf",
  minGSSize = 10,
  maxGSSize = 1000,
  Plot_p.adj = 0.2,
  Plot_Percentage = 10,
  OutputFileName = ""
)

Arguments

InputData

DF with column gene names amd regulatory labels (i.e. sircle clusters). Usually output from SiRCle RCM functions.

geneID

Provide column name for the gene ID in InputData. Needs to match the the gene ID in the pathway file provided

regLabels

Optional: Column name of the column with the regulatory labels in InputData. default: "RegulatoryLabels"

PriorKnowledge

DF including the prior knowledge (PK), ie. pathway file. File must include column "term" with the pathway or TF-regulon name, column "gene" with the gene name and column "Description" with description that will be depicted on the plots.

PKName

Optional: Name of the prior knowledge used, i.e. pathway name such as KEGG default: ""

emptyRegLabel

Optional: Column name of the empty regulatory group default: ""

Plot_FileType

Optional: Output file type for the figures one of: "svg", "pdf", "png" default: "pdf"

minGSSize

Optional: minimum group size in ORA default: 10

maxGSSize

Optional: maximum group size in ORA default: 1000

Plot_p.adj

Optional: q value cutoff from ORA that should be plotted default: 0.2

Plot_Percentage

Optional: Percentage of genes that are detected of a pathway default: 10

OutputFileName

Optional: Output filename default: ""

RemoveBackgroundGenesOptional: If TRUE, genes that fall into background based on the choosen Background method for SiRCle RCM are removed from the universe. default: "TRUE"

Value

DF with the results of the over representation analysis