sircleRCM_2Cond

Computes the regulatory clustering model (RCM) using the SiRCle regulatory rules between two conditions of the same data type.

Usage

sircleRCM_2Cond(
  Cond1_File,
  Cond2_File,
  geneID,
  Cond1ValueCol = "Log2FC",
  Cond1PadjCol = "padj",
  Cond2ValueCol = "Log2FC",
  Cond2PadjCol = "padj",
  Cond1_padj_cutoff = 0.05,
  Cond2_padj_cutoff = 0.05,
  Cond1_FC_cutoff = 1,
  Cond2_FC_cutoff = 0.5,
  backgroundMethod = "C1&C2",
  OutputFileName = "Sircle_RCM.csv"
)

Arguments

geneID

Column name of geneId this MUST BE THE SAME in each of your Input files (we join on this)

Cond1ValueCol

Column name of log fold change in Cond1File

Cond1PadjCol

Column name of p adjusted value in Cond1File

Cond2ValueCol

Column name of log fold change in Cond2File

Cond2PadjCol

Column name of protein p adjusted value in Cond2File

backgroundMethod

Optional: Background method "C1|C2", "C1&C2", "C2", "C1" or "*". For details on the method check https://github.com/ArianeMora/SiRCleR/blob/main/vignettes/SiRCle_RCM_Notebook.Rmd Default="C1&C2"

Cond1File

DF for your data (results from e.g. DeSeq2)

Cond2File

DF for your data (results from e.g. DeSeq2)

Cond1PadjCutoff

Optional: Padjusted cutoff forCond1 data Default=0.05

Cond1LogFCCutoff

Optional: LogFoldchange cutoff for Cond1 data Default=0.5

Cond2FilePadjCutoff

Optional: Padjusted cutoff for Cond2File data Default=0.05

Cond2FileCutoff

Optional: LogFoldchange cutoff for Cond2File data Default=0.5

outputFileName

Optional: Output filename Default=SiRCle_RCM.csv

Value

rcm an instance of the rcm package