sircleRCM_MRP

Computes the regulatory clustering model (RCM) using the SiRCle regulatory rules for DNA-methylation, RNAseq and Proteomics data layers (MRP).

Usage

sircleRCM_MRP(
  methFile,
  rnaFile,
  protFile,
  geneID,
  rnaValueCol = "Log2FC",
  rnaPadjCol = "padj",
  methValueCol = "Diff",
  methPadjCol = "padj",
  proteinValueCol = "Log2FC",
  proteinPadjCol = "padj",
  rna_padj_cutoff = 0.05,
  prot_padj_cutoff = 0.05,
  meth_padj_cutoff = 0.05,
  rna_FC_cutoff = 1,
  prot_FC_cutoff = 0.5,
  meth_Diff_cutoff = 0.1,
  backgroundMethod,
  OutputFileName = "Sircle_RCM"
)

Arguments

methFile

DF for your DNA methylation data (results from differential methylation analysis)

rnaFile

DF for your RNAseq data (results from DeSeq2 and also your normalised expression counts)

geneID

Column name of geneId this MUST BE THE SAME in each of your protein, RNAseq and DNAmethylation files (we join on this)

rnaValueCol

Column name of RNA value in rnaFile

rnaPadjCol

Column name of RNA p adjusted value in rnaFile

methValueCol

Column name of Methylation difference value in methFile

methPadjCol

Column name of Methylation p adjusted value in methFile

proteinValueCol

Column name of protein log fold change in proteinFile

proteinPadjCol

Column name of protein p adjusted value in proteinFile

backgroundMethod

Optional: Background methods: "P|(M&R)", "P|M|R", "P|R", "P&R", "P&M&R", "(P&M)|(P&R)|(M&R)", "(P&M)|(P&R)" or "*". For details on the methods, please check https://github.com/ArianeMora/SiRCleR/blob/main/vignettes/SiRCle_RCM_Notebook.Rmd Default="P|(M&R)"

proteinFile

DF of you Protein data (results from DeSeq2 and also your normalised expression counts)

rnaPadjCutoff

Optional: Padjusted cutoff for RNAseq data Default=0.05

rnaLogFCCutoff

Optional: LogFoldchange cutoff for RNAseq data Default=0.5

proteinPadjCutoff

Optional: Padjusted cutoff for Protein data Default=0.05

proteinValueCutoff

Optional: LogFoldchange cutoff for Protein data Default=0.3

methPadjCutoff

Optional: Padjusted cutoff for DNA methylation data Default=0.05

methDiffCutoff

Optional: DNA Methylation difference cutoff for DNA methylation Default=10

outputFileName

Optional: Output filename Default="SiRCle_RCM"

Value

DF with input data combined by geneID and SiRCle Clusters